# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zima Display Automation API
  version: 0.4.0
servers:
  - url: http://zimaos-host/zima-display/api/v1
security:
  - bearerAuth: []
paths:
...
  /presentations/text:
    post:
      summary: Create a text presentation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, pages]
              properties:
                title: {type: string}
                source: {type: string}
                pages:
                  type: array
                  maxItems: 500
                  items: {type: string}
      responses:
        "201": {description: Presentation created}
  /presentations:
    post:
      summary: Upload a ZIP containing presentation page images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archive]
              properties:
                archive: {type: string, format: binary}
                title: {type: string}
                source: {type: string}
      responses:
        "201": {description: Presentation created}
...
  /action:
    post:
      summary: Run a display or media action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [mode, play, pause, resume, stop, seek, volume, next, previous]
                mode:
                  type: string
                  enum: [dashboard, canvas, clock, black, terminal]
                path: {type: string}
                paths:
                  type: array
                  items: {type: string}
                value: {type: number}
      responses:
        "200": {description: Action completed}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer